# [PARQUET-1542] Merge multiple I/O to one time I/O in method readFooter