Currently in method readFooter of class ParquetFileReader, it will invoke method converter.readParquetMetadata. To decode file metadata,  there may be quite a number of times of I/O processing underlying, which may heavily slow down parquet read performance.

A better way is to read all the footer bytes in a single I/O, since we have already got footer length before, and pass down all the bytes to underlying.

And we have a test in our environment, this change may bring  50 percent improvement for some cases.