@nandorKollar 2019-02-27
There are test failures, however they look unrelated, it looks like Twitter maven repo is down, and we pull something from there. Travis should be retriggered once it is up again.