@gszadovszky 2019-02-27
I'm afraid I don't understand why your change would boost the reading. As far as I can see thrift reads all the bytes of the footer only that in some cases not the whole thrift structure is created. Thrift does not seem to do any further seeks or skipping bytes in the underlying input stream.
Are you sure this change made your system working faster? Could you write a benchmark that proves this change would increase the performance in your scenario and would not decrease it significantly in any other ones?