@wangshisan 2019-02-27
Underlie of method converter.readParquetMetadata, it will read from the InputStream and deserialize to a FileMetaData object. During serialization, it will read all the filed bytes one by one, you can find this logic in method FileMetaData.read(), unfortunately, the InputStream is without any buffer, that means for each field, there will be at least one I/O, this is pretty time consuming especially for HDFS.