@gszadovszky 2019-02-27
I see, thanks for describing.
Could you please provide some benchmarks for different scenarios (by simulating remote reads, reading local files etc.). It would be nice to see that this change really improves the reading in specific cases (if not all of them) and does not have a significant penalty for the others.