@wangshisan 2019-09-21
@gszadovszky Sorry, I'm not quite clear, do you mean I need add a benchmark in Parquet-Benchmarks module, or some benchmark result in our Product environment?