@gszadovszky 2019-09-23
The best thing would be adding some benchmarks into the parquet-benchmarks module. However, I am not sure how can we properly simulate HDFS. On local disks, it probably won't have any performance benefits because the OS would pre-read the block to memory before giving it to the java reader.
I don't know if HDFS would do similar tricks on remote reads or it would stream the data through the wire. That's why I am interested in some benchmarks. Currently, I am not convinced this change would boost the footer reading.
If you are able to do some benchmarks in your environment with HDFS remote and local reads with and without this fix and provide the results, it would be great. :)