@wangshisan 2019-09-23
I made a benchmark on two of our production tables.

- Table a, 66 columns:

Run | Before(ms) | After(ms)
-- | -- | --
Run 1 | 65 | 40
Run 2 | 67 | 43
Run 3 | 59 | 43
Run 4 | 65 | 46
Run 5 | 60 | 39



- Table b, 155 columns.

Run | Before(ms) | After(ms)
-- | -- | --
Run 1 | 85 | 52
Run 2 | 99 | 57
Run 3 | 83 | 56
Run 4 | 91 | 53
Run 5 | 87 | 52