@gszadovszky 2019-09-23
Thanks a lot, @wangshisan.
Do you know if these all were remote reads? It would be surprising to me if your change would boost local reads.
Anyway, I'll approve this change.